misc/src/com/unitedjiga/common/parsing/Token.java
/*
 * The MIT License
 *
 * Copyright 2020 Junji Mikami.
 *
 * Permission is hereby granted, free of charge, to any person obtaining a copy
 * of this software and associated documentation files (the "Software"), to deal
 * in the Software without restriction, including without limitation the rights
 * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
 * copies of the Software, and to permit persons to whom the Software is
 * furnished to do so, subject to the following conditions:
 *
 * The above copyright notice and this permission notice shall be included in
 * all copies or substantial portions of the Software.
 *
 * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
 * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
 * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
 * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
 * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 * THE SOFTWARE.
 */
package com.unitedjiga.common.parsing;

/**
 *
 * @author Junji Mikami
 */
public interface Token extends TerminalSymbol {

    /**
     * 
     * @param value
     * @return
     */
    static Token of(String value) {
        return new Token() {
            private String s = String.valueOf(value);
            @Override
            public String getValue() {
                return s;
            }
            @Override
            public String toString() {
                return s;
            }
        };
    }

    /**
     * 
     * @param t1
     * @param t2
     * @return
     */
    static Token concat(Token t1, Token t2) {
        return of(t1.getValue() + t2.getValue());
    }

    @Override
    String toString();
}
